requirements.txt
openai
whisper
streamlit
faiss-cpu
langchain
pymupdf
python-docx
youtube-transcript-api
pytube
